+++ CHANGELOG.md
@@ -9,11 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - #889, Allow more than two conditions in a single and/or - @steve-chavez
 - #883, Binary output support for RPC - @steve-chavez
+- #885, Postgres COMMENTs on SCHEMA/TABLE/COLUMN are used for OpenAPI - @ldesgoui
 
 ### Fixed
 
 - #877, Base64 secret read from a file ending with a newline - @eric-brechemier
 - #896, Boolean env var interpolation in config file - @begriffs
+- #885, OpenAPI repetition reduced by using more definitions- @ldesgoui
 
 ## [0.4.2.0] - 2017-06-11